requirements.txt
langchain
langchain-openai
langchain-anthropic
python-dotenv
langchain-community
pydantic
duckduckgo-search
ddgs
langchain-deepseek